correlation-analysis/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pytz>=2024.1
mcp>=1.0.0
